single tag watcher split
